clean up and add comments

+++ api/controllers/authController.js
@@ -3,13 +3,17 @@ const authController = {
         const { MockCache, authService } = objectRepository;
         return async (req, res, next) => {
             const { username, password } = req.body;
+
+            if (typeof username === "undefined" || typeof password === "undefined")
+                return res.status(400).json({ message: "Bad request!" });
+
             try {
                 const user = await authService.getUser(objectRepository)(username, password);
                 if (!user) return res.status(400).json({ message: "Invalid login credentials!" });
 
                 const accessToken = await authService.getAccessToken(objectRepository)(user);
                 const refreshToken = await authService.getRefreshToken(objectRepository)(user);
-                MockCache.refreshTokens.push(refreshToken);
+                MockCache.refreshTokens.push(refreshToken); // TODO: change MockCache to Redis
 
                 return res.status(200).json({ username: user.name, isAdmin: user.isAdmin, accessToken, refreshToken });
             } catch (err) {
@@ -22,6 +26,7 @@ const authController = {
         const { authService } = objectRepository;
         return async (req, res, next) => {
             const refreshToken = req.body.token;
+            if (typeof refreshToken === "undefined") return res.status(400).json({ message: "Missing refresh token!" });
 
             const destroyToken = await authService.destroyRefreshToken(objectRepository)(refreshToken);
             if (!destroyToken) return res.status(400).json({ message: "Failed to log out!" });
@@ -35,13 +40,13 @@ const authController = {
             const refreshToken = req.body.token;
             if (!refreshToken) return res.status(401).json({ message: "You aren't authenticated!" });
 
-            if (!MockCache.refreshTokens.includes(refreshToken)) return res.status(403).json({ message: "Invalid refresh token!" });
+            if (!MockCache.refreshTokens.includes(refreshToken)) return res.status(403).json({ message: "Invalid refresh token!" }); // TODO: change MockCache to Redis
 
             try {
-                const { newAccessToken, newRefreshToken } = await authService.refreshExpiringRefreshToken(objectRepository)(refreshToken);
+                const { newAccessToken, newRefreshToken } = await authService.refreshExpiringTokens(objectRepository)(refreshToken);
                 return res.status(200).json({ accessToken: newAccessToken, refreshToken: newRefreshToken });
             } catch (err) {
-                const statusCode = err.statusCode || 400;
+                const statusCode = err.statusCode || 400; // TODO: custom errors yet to be implemented
                 return res.status(statusCode).json({ message: err.message });
             }
         }
